# custom-api-agent/openapi.yaml
openapi: 3.0.3
info:
  title: Booking Agent API
  version: "1.0.0"
servers:
  - url: http://localhost:8085
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /policy:
    get:
      summary: Get the current cancellation policy
      responses:
        "200":
          description: Cancellation policy text
          content:
            application/json:
              schema:
                type: object
                properties:
                  cancellation_policy:
                    type: string
  /bookings:
    get:
      summary: List bookings for a user
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCreateRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingCreateResponse"
  /bookings/{booking_id}:
    get:
      summary: Get a booking by id
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
    put:
      summary: Update a booking
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdateRequest"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingMutationResponse"
    delete:
      summary: Cancel a booking
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingMutationResponse"
components:
  schemas:
    Booking:
      type: object
      properties:
        booking_id:
          type: string
        confirmation_number:
          type: string
        user_id:
          type: string
        hotel_name:
          type: string
        check_in_date:
          type: string
        check_out_date:
          type: string
        number_of_guests:
          type: integer
        primary_guest:
          type: object
        special_requests:
          type: string
        booking_status:
          type: string
        booking_date:
          type: string
        updated_at:
          type: string
        cancelled_at:
          type: string
    BookingCreateRequest:
      type: object
      properties:
        user_id:
          type: string
        check_in_date:
          type: string
        check_out_date:
          type: string
        number_of_guests:
          type: integer
        primary_guest:
          type: object
        special_requests:
          type: string
      required:
        - check_in_date
        - check_out_date
        - number_of_guests
        - primary_guest
    BookingUpdateRequest:
      type: object
      properties:
        user_id:
          type: string
        check_in_date:
          type: string
        check_out_date:
          type: string
        number_of_guests:
          type: integer
        primary_guest:
          type: object
        special_requests:
          type: string
    BookingCreateResponse:
      type: object
      properties:
        booking_id:
          type: string
        confirmation_number:
          type: string
        confirmation_message:
          type: string
        cancellation_policy:
          type: string
        booking_details:
          $ref: "#/components/schemas/Booking"
    BookingMutationResponse:
      type: object
      properties:
        message:
          type: string
        booking_details:
          $ref: "#/components/schemas/Booking"